Make Bool a data type. (#172)
This PR makes Bool into a built-in `data` type. Now we can pattern match
on Bool! This lets us remove four primitives. (In theory we can also
remove the `if` and `while` primitives as well, but we don't yet have a
replacement for their built-in optimizations, so better to leave them.)