Overloading based on top stack type. (#178)
Kind of a big change. This PR makes it so that if a word doesn't exist
[and the word doesn't start with uppercase or underscore], then the
elaborator looks at the top type on the stack and prepends the name of
the type constructor to the word [also adding a `.` if the word started
with a letter].

For example, `+` is not defined, but `Int+` is. So when you do `1 2 +`
the elaborator replaces the `+` with `Int+` because the top stack item
is an `Int` at that point. Thus we could have another type, say `Float`,
implement `Float+`, and then we could reuse `+` for `Float+` or `Int+`
depending on the context. Voila, operator overloading. (Note we do not
actually have a `Float` type at this point.)

Another example, `map` is not defined, but `List.map` and `Maybe.map`
are defined. When you have a list, and you do `map(1+)`, the elaborator
replaces it with `List.map(1+)`. When you have a maybe, the elaborator
replaces it with `Maybe.map(1 +)`. Note that, unlike typeclasses or
traits, there isn't a generic `map`.

I also added a special case for `Mut`, so the elaborator doesn't see it
when determining the type constructor. So e.g., even though `Stack.pop!`
has type `Mut(Stack(t)) -- t`, if you use `pop!` the elaborator can
still find `Stack.pop!`, ignoring the `Mut` part of the type.

Besides implementing the overloading, this PR also uses it clean up the
prelude and data structures. Lots of good use cases for it, e.g.
relational operators, casting between integer types, common higher-order
operations like `map`, etc. I also started using it in the compiler
proper, but that change isn't done. The compiler is kinda big . . .