Add linear resource types to Mirth. (#182)
This PR adds a resource stack alongside the usual value stack. Resources
are linear, so they cannot be `dup`ed or `drop`ed. They can be used to
represent mutable state in a program. They can also be used as
capabilities for external interaction (e.g. file handles) requiring
explicit creation, destruction, and sequencing of all the operations in
between.

What is implemented in this PR:

1. TYPES: Resources are incorporated into `StackType` with the `STWith`
constructor. This means that `StackType` represents both the value stack
and the resource stack. There's special logic inside `StackType`
unification so that the `STCons` and `STWith` stack type constructors
commute against each other as needed, which makes resources and values
act as independent stacks for the purposes of type checking.

- When printing out a `StackType`, the value types are printed before
all the resource types, giving a normalized representation. The internal
representation via `STWith` and `STCons` should not matter to the
language user.

2. PARSING: Resources can show up in type signatures, whenever a stack
type is written. Resource types all start with a `+` to distinguish them
from value types. Defined resource types have an uppercase letter
following the `+`, e.g. `+C99` is a resource defined in the C99 codegen
module, while resource type variables have a lowercase letter, e.g.
`+a`. Resources can be placed anywhere along the stack, it should not
matter if you put them before or after the values, but there is a
preference to use the normalized representation (except: see pattern
matching unification bug below).

3. ALGEBRAIC RESOURCES: The language user can define their own resource
types via `data`, similar to defining a value type. These algebraic
resource types can have multiple constructors, and these constructors
can take both value types and resource types as input. In this PR, this
has been used to define a `+C99` resource that wraps over the shared
state needed in the C99 codegen.

4. OVERLOADING: The top resource type on the stack can be used for name
overloading, either by itself (e.g. `+C99.line` can be invoked as `line`
in a `+C99` context), or together with the top value type on the stack
(e.g. the word `Str+C99.put` with type `Str +C99 -- +C99` can be invoked
as `put` in the appropriate context). In this PR, this has been used to
significantly clean up the C99 codegen.

Small issues that need to be addressed very soon:

- Pattern matching unification bug: When pattern matching with an
underscore `_`, the presence of a resource type on the top of the stack
causes a unification error. Moving the resource type below the top of
the value stack eliminates the unification error, so I'm sure this is a
unification bug relating to the `STCons` and `STWith` commutation.

- Smuggling resources into values: Right now there is nothing preventing
the language user from defining a value type via `data` with a
constructor that takes a resource as input. This needs to be prevented,
because the value type can be used to `dup` and `drop`, leading to
non-linear resource usage.

Broader issues to tackle later:

- Kind checking & higher-order resources: Right now, only value types
can be used as arguments to type constructors. Resource types cannot be
used, and so higher-order resource types are possible, but they can only
take value types as an argument. :-( This issue goes hand in hand with
implementing kind checking, which is completely absent from Mirth at
present because there weren't any meaningful kind distinctions until
now.

- Reusing value types as resources: Right now resource types and value
types have different names, so e.g. there isn't a way to reuse `Maybe`
or `Either` as a resource type (not that you could use it meaningfully,
due to the previous point), without defining also a `+Maybe` and
`+Either`. At this point, I don't know if reuse of the definition of
value types is reasonable, or if explicit redefinition is better (in
some cases, or in all cases). It's something to investigate.

- Records: `data` resources are useful for modelling resources with
multiple possible states, e.g. a ready state and an error state, but
more commonly we would like to work with resources that represent
multiple parts that can be used independently, for which these algebraic
resource types (ARTs!) are a bit cumbersome to use because they rely
exclusively on positional notation. Mirth is long overdue some kind of
record type construction (tables tend to fill the role of records in the
compiler), and resources make that need more dire.

- Using resources consistently: Up until now, Mirth has had a very
relaxed approach to mutability and side effects. This was crucial to
getting things up and running. But, resources are a way to model both
mutable state and many "side effects" in a referentially transparent
way. It would be good to move the existing compiler onto resources, as
has already been done with the bulk of the C99 codegen in this PR, and
perhaps in doing so find out where the weaknesses in using resources
lie. But this is a big undertaking, so expect it to happen in pieces
over a span of time.

[self-promo] If you enjoyed reading this PR summary, please consider
supporting my [Patreon](https://patreon.com/typeswitch), and/or consider
hiring me to work on your compiler project (my email is typeswitch at
protonmail dot com).