Remove prim-unsafe-cast, replace legitimate uses with OP_COERCE. (#193)
This PR removes prim-unsafe-cast and replaces the remaining uses of it.
As part of this, it introduces new compiler-generated `T.tag` and
`T.from-tag-unsafe` whenever `T` happens to be an enum type.

The problem with prim-unsafe-cast is that it exposes the underlying
representation of values at runtime. By replacing it with an operator
OP_COERCE, we can control where and when it is used. For now there isn't
much control over usage of OP_COERCE, but it's only used in
compiler-generated words.