Introduce namespaces (#198)
This PR implements namespaces. This means we can now have two words with
the same name, as long as they're defined in different namespaces.
Namespaces exist per package, per module, and per type constructor. As
part of this PR we also introduce a simple package-and-module hierarchy
for source code, and some syntax for qualified and relative names.

### Package and Modules

Source code has been reorganized into a simple two-level hierarchy of
packages and modules, for example:

```
src/
   package1/
       module1.mth
       module2.mth
       ...
   package2/
       module1.mth
       module2.mth
       ...
   package3/
       ...
   ...   
```
In other words, a package is a directory that containing `.mth` files as
module. At the moment this is a fixed two-level hierarchy, with all
packages kept in the `src` folder. The current set of packages are:

- `std`: common data structures and algorithms
- `posix`: I/O and filesystem support via the POSIX API
- `args`: the argument parser developed by @casuallyblue 
- `mirth`: the mirth compiler itself
- `snake`: the "snake clone" demo, which includes some incomplete SDL2
bindings

### Namespaces and Qualified Names

Each module has its own namespace, so you can now define the same word
in two different modules, without a problem. For example, suppose we
defined a word `foo` in both `package1.module1` and `package2.module2`.
Then suppose we imported both `package1.module1` and `package2.module2`
in a third module, The name `foo` will be ambiguous, but we can use the
fully qualified names `package1.module1.foo` and `package2.module2.foo`
to disambiguate between these definitions whenever it becomes necessary.

Besides module namespaces, there's also type namespaces. When you have a
type constructor, e.g. `Int` or `List`, it has a namespace associated
with it. We can define words within that namespace, and they'll be
invoked when that type is on the top of the stack, e.g. if we define
`Int.show` and `List.show`:

```
   def(Int.show, Int -- Str, ...)
   def(List.show, List(t) -- Str, ...)
```

Then whenever we use the word `show`, if the top of the stack is an
`Int` the compiler will use `Int.show`, and if the top of the stack is a
`List` then the compiler will use `List.show`. However, if there is also
a word called `show` in scope, this will be ambiguous. So in that case
the compiler requires that you to disambiguate. There's a few ways to do
this:

1. If the relevant type is on top of the stack, and you want to invoke
the word from that type's namespace, then you can use a "relative name",
which would simply be `.show` in this case. This will use the type on
top of the stack to pick the correct definition to use, while ignoring
any other `show` in scope.
2. If the relevant type is in scope, you can use a type-qualified name,
e.g. `Int.show` or `List.show`, to use the word from the type's
namespace. This not only resolves ambiguity, but also provides some type
information to the reader and the compiler.
3. Failing these, you can always use the fully qualified name of a word,
e.g. `package1.module1.show` or `std.prim.Int.show`, to get a particular
definition. Every definition has a unique fully qualified name, so this
method always works.

Besides the top value type on the stack, the compiler also uses the top
resource type on the stack, to do disambiguation. So you can have words
like `+Output.put` which can be invoked as `put` as long as the
`+Output` resource is on top of the resource stack.

This system replaces the previous method of type-directed overloading,
which involved prefixing a name with the input types. In particular the
new system is actually less powerful in one way: there's no longer a way
to dispatch based on both the top value type and the top resource type,
But in return, we have a system that is more regular, more well behaved,
and it has the predictable fallback of just adding more qualifications
as needed.

Finally, one more thing worth mentioning is that names that begin with
an uppercase letter (i.e. "constructor names") behave slightly
differently with respect to overloading: the plain unqualified version,
e.g. `NONE`, does not participate in type-directed overloading. That's
because the type at the top of the stack isn't often a relevant factor
for constructors, unlike the output type*. Constructors are considered
part of the type's namespace for the type that they construct, although
they can also be invoked without qualification as long as the type is in
scope. So we can use `NONE`, `Maybe.NONE`, or `std.maybe.Maybe.NONE` to
refer to this same constructor.

(*At the moment, the output type is not used for type-directed name
resolution of constructors. I plan to investigate whether this kind of
type-directed name resolution is a feasible thing to implement in mirth,
& find out whether it leads to better type inference and nicer code.)