Implement Stack Labels (#210)
This PR implements stack labels. This gives us the ability to attach
labels to any value on the stack. This has the effect of pushing the
value onto a private stack for that label. The main use for this is to
implement keyword arguments, as opposed to relying solely on positional
arguments.

For example, we can define a word that takes its arguments using labels,
```mirth
def(my-word, foo:Int bar:Int baz:Str -- Str, ...)
``` 
Then to use this word we must push its arguments on the corresponding
label stacks, using `>foo`, `>bar`, and `>baz` (read: "to foo", "to
bar", and "to baz"),
```mirth
10 >foo 20 >bar "hello world" >baz my-word
```
The advantage of labels is that the order does not matter, so the
following lines are all equivalent to the above:
```
20 >bar 10 >foo "hello world" >baz my-word
"hello world" >baz 20 >bar 10 >foo my-word
10 >foo "hello world" >baz 20 >bar my-word
```
When pushing values to a label, they leave the regular stack. So
labelled stack values do not interfere with the rest of the stack, i.e.
the positional stack. We could write the above like so:
```
10 20 >bar >foo "hello world" >baz my-word
"hello world" 20 10 >foo >bar >baz my-word
```
We can use labels to implement our own `dip`, for example:
```
10 20 "hello" >x + x>
```
This is the same as `30 "hello"` because the `>x` hides the `"hello"`
value in the `x` label stack, and then `x>` restores it. This is similar
to `>r` and `r>` in Forth, except that label stacks to not interfere
with the return stack, so it's safe to write a word that places some
value on a label stack and then returns, e.g.
```
def(label-tuff, Int Str -- foo:Int bar:Str, >bar >foo)
```
So labels are compositional (and concatenative) in a way `>r` and `r>`
sometimes fail to be. Also, label names do not need to be declared in
advance, you can just use them.

This PR implements the above, adds some uses for labels (when dealing
with command-line arguments) and a test. This PR does not implement
label arguments for `data` constructors. That will be future work.