Generate unpacking and label projection functions for single-constructor data types automatically. (#213)
This PR changes `data` to make it generate a few functions in the case
where the `data` type has a single constructor. Suppose we have a data
type called `Foo` and it has a single constructor `FOO`. Then the
compiler will generate:

- a method `Foo./FOO` that undoes the action of the constructor. It is
defined as `match(FOO -> )`

- for each labelled input to the `FOO` constructor, for example
`bar:Bar`, it generates a method `Foo.bar` of type `Foo -- Bar` that
fetches that field. This is defined as `/FOO bar> dup >bar dip(FOO
drop)`

For the latter, if the data is actually a resource, it doesn't do the
`drop`, so the method `+Foo.bar` will have type `+Foo -- +Foo Bar` and
is defined as `/FOO bar> dup >bar dip(FOO)`.