Make colon a right-associative application operator -- lens syntax. (#215)
This PR implements reserves colon (`:`) as a special token, and turns it
into a right-associative application operator, as an alternative for
parentheses in many situations.

This new syntax can be used in place of unary application, whenever the
argument consists of a single atom (including the atom's arguments). For
example `foo(bar)` can be replaced with `foo:bar`. The syntax is
right-associative, so `foo(bar(baz))` can be replaced with
`foo:bar:baz`.

The primary motivation for this syntax is to be able to compose multiple
higher-order words without creating a mess of parentheses. In
particular, this syntax should be useful for defining and using
functions to access deeply nested fields. Hence calling it "lens
syntax".

Examples of the syntax:
- `foo:bar` is equivalent to `foo(bar)`
- `foo:bar(10)` is equivalent to `foo(bar(10))`
- `foo:bar:10` is equivalent to `foo(bar(10))`
- `foo:bar:baz` is equivalent to `foo(bar(baz))`
- `foo:bar:baz(1 2 3)` is equivalent to `foo(bar(baz(1 2 3)))`

That said, there is one place where these two are not equivalent, and
it's in type signatures. A `foo:Int` in a type signature indicates a
stack label (see #210). It is not a type variable application*
`foo(Int)`. Internally the `:` syntax and the traditional `()` syntax
are stored as different kinds of application syntax, so a distinction
can be made as needed.

Implementation detail: The `:` syntax introduces two new token types:
one for opening the colon application (called `TOKEN_LCOLON`) and one
for closing the colon application (called `TOKEN_RCOLON`). These are
placed before and after the argument respectively, where the equivalent
`TOKEN_LPAREN` and `TOKEN_RPAREN` would have been if the traditional
application syntax had been used instead. While `TOKEN_LCOLON` tokens
correspond to the colon character `:` in the source code, the
`TOKEN_RCOLON` tokens don't correspond to any lexical token in the
source code. They are generated by the lexer as virtual tokens. These
make the next stages of the compiler easier and more uniform, by making
`:` application more similar to `()` application. The tokenized language
remains visibly pushdown.

(*-type variable application is not currently supported)