Allow for multiple words with the same name & different arities. (#216)
This PR makes word arity part of the internal qualified name of the
word. As a result, we can have multiple words with the same name, as
long as they have different arities.

Consequences:
- Currying is no longer automatic.
- When setting an `alias` for a higher-order word, you have to indicate
the number of parameters as part of the alias definition. This is done
by supplying a parameter list. Other than the arity, the argument list
is ignored.

[Also incidentally I added `+World.put` and `+World.line`]