Shorten generated names. Inline specialized words and enum constructors. Avoid generating unnecessary tag words. (#228)
This PR enables inlining for specialized higher order words (unless they
are recursive). It also adds inlining for enumeration constructors and
patterns, and prevents generating code for tags that aren't needed. The
overall goal was to improve the quality of the generated code and to
reduce its size.

One neat result is that it's now faster to compile using -O0 on my
machine. This will make optimization even more important.