Implement name resolution for constructors in pattern matching. (#254)
This PR brings good name resolution to constructors in `match`
statements. You can now have constructors of the same name without fear,
because `match` will prefer the constructor in the type's namespace. If
there is still ambiguity (e.g. if there's both a value type and a
resource type with the same constructor name available to match at the
same time -- see the example in the test) you can use a qualified name
to resolve it.

Closes #227