Add a check for inline recursion. (#255)
Adds a check that a word marked as inline cannot be directly recursive
(or indirectly recursive through other inline words). If it is found to
be recursive, a warning is issued and the word is marked as not inline.

Fixes #249 (importing std.prelude as inline will just generate a bunch
of warnings now).