Use cons list in std.list (#269)
This PR simplifies std.list to be just a cons list. This was after lots
of experimentation and testing to see what would be nice to work with
and be fast. Despite their simplicity (or maybe because of it), cons
lists were the fastest, saving several seconds over unbalanced and
balanced cat lists with or without len, and finger trees (as measured by
how long it takes me to compile the compiler, by running `time make`).

A very rough summary of the experiment results (mostly from memory):

- simple cons lists: 34 seconds
- cat lists unbalanced: 35 s
- cat lists balanced with cached len: 37 s
- cat lists balanced with no cached len: 48 s
- finger trees (non-lazy): 47 s