Add unit tests for std.list (#270)
This PR adds a module (`mirth-tests.std-list`) with extensive unit tests
for `std.list`. In addition:

- implements a unit testing library, `std.test`.

- changes the specializer to only specialize higher order words that
have been marked as inline. this change actually slows down the compiler
a little ... but it speeds up the testing significantly, reducing the
generated C size by almost 90%. i guess it's because the unit testing
library uses a lot of higher order functions to make it look nice, and
the specializer is all too eager to inline everything, resulting in a
huge generated code.

- introduces `Int.to:` and `Int.from:` as alternatives to `Int.range`
taking one of the bounds as a parameter. So you can write `1 to: 4` to
get the list, `1 2 3 4`, or `5 from: 0` to get the list `0 1 2 3 4 5`.
(these will be moved out of std.list at some point, when they no longer
produce lists.)

- closes #242