Add support for resource labels. (#276)
This PR adds support for using stack labels with resources. List of
changes:

- Labels can be placed on resources in stack types, e.g. `lexer: +Lexer`
or `my-resource: +a`

- You can "push" and "pop" resources onto labels using `+>foo` and
`foo>+`. In particular:

  -  `+>foo` pushes a resource onto the stack label `foo`
  - `foo>+` pops a resource from the stack label `foo`

- As usual, stack labels behave like a separate stacks for each label.
But note that it is *one stack per label*, there is no separation of
resources and non-resources in a stack label, it's all combined into one
stack. So for example, you cannot push a value to a stack label, then
push a resource to the same stack label, and then pop the value from
that stack label. That's because the resource is currently on top. You
need to pop the resource first, then you can pop the value.

- You can use labelled resources in a data definition, e.g.:

  ```
  data(+Output, +Output ->
      file: +File
      buffer: +Buffer)
  ```

- If a data definition has a single constructor, like the example above,
then the compiler generates words to access the resources contained
within automatically. Using the example above we have:

- `+Output.file(f)` and `+Output.buffer(f)` will run the function `f`
using the resource within:
    ```
+Output.file(f) :: (*a +File -- *b +File ) *a +Output -- *b +Output
+Output.buffer(f) :: (*a +Buffer -- *b +Buffer) *a +Output -- *b +Output
    ```
    
- `+Output.file` and `+Output.buffer` will split the `+Output` resource
in two, giving access to the underlying resource:
    ```
    +Output.file   :: +Output -- +File   +Output/file
    +Output.buffer :: +Output -- +Buffer +Output/buffer
    ```
Note that `+Output/file` and `+Output/buffer` are internal types that
you can't write down directly in the surface language. They represent an
`+Output` resource that has had a "hole" punched in it, where `file` and
`buffer` used to be, respectively.
    
- `+Output.file!` and `+Output.buffer!` will put the `+Output` resource
back together, undoing the actions of `+Output.file` and
`+Output.buffer`:
    ```
    +Output.file!   :: +File   +Output/file   -- +Output
    +Output.buffer! :: +Buffer +Output/buffer -- +Output
    ```
    
- Note that `+Output.file(f)` and `+Output.buffer(f)` are actually
defined in terms of the other two:
  
    ```
    +Output.file(f)   = +Output.file rdip(f) +Output.file!
    +Output.buffer(f) = +Output.file rdip(f) +Output.buffer!
    ```
    
- Lastly, I tweaked the syntactic sugar for accessing labels without
pushing or popping. Instead of `^foo` to access a label, you now have:
  
  - `!foo` to set the top value: `!foo = foo> drop >foo`
  - `@foo` to get the top value: `@foo = foo> dup >foo`
- `@foo(f)` to run `f` on the top value or resource: `@foo(f) = foo> f
>foo` or `foo>+ f +>foo` depending on the type.
  
Of these, the only one that will work with resources is `@foo(f)`, since
it doesn't require duping or dropping a resource.