Package resolution: Derive package path from module path, and have a default package search path. (#282)
This PR implements two main changes:

- When loading a module via a module path without knowing a package path
(e.g. because the module path is passed in as the command-line argument
to the mirth compiler), then infer the package path from the module
path. e.g. if loading module `foo.bar`, from `my/directory/bar.mth`, it
will infer that package `foo`'s path is `my/directory`.

- When loading a module via its name, and you don't know where the
package is, it will look for a package in a predefined list of
directories -- the "package search path". Right now, the search path is
hardcoded to be just the "lib" directory, but it should be easy to make
package search path configurable in the future (e.g. adding a `-P`
option).

Together these eliminate all the necessary uses of `-p` package flags in
the mirth repo. Of course `-p` still works to specify a package path
manually ... we just don't need to do it.

Also this PR fixes issue #199.