Add +Mirth resource, prefer show; over show ; (#284)
This started out as a small PR to change `show` and `trace` words into
`show;` . This led to me trying to remove `trace` where it really
shouldn't be. Which led to me creating a resource (`+Mirth`) to collect
diagnostics while the compiler is running and display them at the end.
Which also led to me redo-ing pieces of the "lazy" computation in order
to thread in this new resource, which led to the beginnings of a
propagator-like API within `+Mirth`. And so the small PR turned into a
massive PR, but it's good, because I wanted to make those changes all
along.