Use custom tag values in sdl bindings. (#306)
Use custom tag values in sdl bindings. The biggest benefactor of this is
the SDL_EventType, which you can now pattern match on instead of using
writing `if` statements.

Also discovered the reason why controller input was ignored in the past
-- it as reading the wrong field of the SDL_Event data structure. (Now
controller input will still be ignored in snake, but that's because it
no longer tries to open a controller.)