Make package search paths part of the +Mirth resource and add a CLI argument to add to it (#311)
Replaces the existing PACKAGE_SEARCH_PATH variable with a field in the
+Mirth resource and adds the `-P` command line flag to add items to it,
without any package search path flags it defaults to `lib`. This also
replaces the couple `data` definitions related to +Mirth with the struct
syntax sugar, but that's easy to undo separately from the rest of the
changes if not wanted.