Keep track of values in C99 codegen (#316)
This PR changes the C99 codegen significantly -- it adds a "stack cache"
that keeps track of what values are in each stack/stack label, so as to
avoid unnecessary pushing and popping. Taking full advantage of this, we
also introduced a suitable (internal) API for words so that argument
passing & return values don't use the stack. Altogether this cuts out
close to 90% of all global stack pushes and pops, with the remaining
being caused by non-inlined closures, weird type signatures (like
panic!), and poor or insufficient type information in the codegen.