Removing a bunch of panic! calls. (#317)
The focus in this PR was removing a lot of panic! calls by either
strengthening types or exposing errors in return values. The biggest
changes from this were:

* getting rid of `Maybe.unwrap` (though we still have `Maybe.unwrap/1`
where you provide a fallback yourself -- where you can panic! if that is
really the best thing to do) and similar unconditional unwrapping
functions that panic.

* removing any integer conversion functions (like `Int.>Nat`) that panic
in favor of ones that either make you handle overflow, or ones that
truncate or clamp instead of panicking.

* `<data>.tag` and `<data>.from-tag-unsafe` were renamed to
`<data>.cons-value` and `<data>.from-cons-value-unsafe`, because .tag is
a bit too useful.