Add default parameters for higher-order words. (#323)
This PR adds the ability to define a default for parameters to
higher-order words. The default is evaluated in the caller's
environment, so this can be used to pass a type-aware equality, and
things like that, automatically.

Example:

```
def foo (f { bar } ) { f f }
def bar { ... }
```
Here, if we call `foo` with no arguments, this is the same as calling
`foo(bar)`.

More useful example, defining a higher-order equality word on generic
types:

```
def Maybe.== (eq {==})  {
   { None -> none? }
   { Some -> swap match {
      { None -> drop False }
      { Some -> swap eq }
   }
}
```

Now when we call `==` on a Maybe value, the mirth compiler fills in the
default argument so it is equivalent to writing `==(==)`. Therefore we
can call `==` on Maybe to test equality for a bunch of different types,
without having to pass in the function explicitly.

Things that can be improved: the diagnostics that arise from errors when
a default parameter doesn't fit. The name resolution for words that have
default parameters could also be tighter.