Refactor Lexer, fix chained label token lexer bug, add !foo/1 label helper, rename some Maybe words. (#339)
This started as a lexer bugfix (chained label tokens were broken) that
turned into a refactoring, that resulted in some changes to Maybe words
and a new label helper function . . .

- Refactored the lexer to use the new `data` syntax for the most part.
- Fixed a bug that meant label helper functions like `@foo/1` were not
chaining correctly with colon. With this fix we can now write
`@x:@y:swap` for example, to the top of the `x` and `y` labels.
- Renamed `Maybe.if-some` and `Maybe.or-some` to `Maybe.if?` and
`Maybe.or?` respectively. Added `Maybe.then?`, `Maybe.else?`, and
`Maybe.and?` functions as well, which are "Maybe" counterparts to the
respective boolean-flavored words.
- Added a new label helper syntax, `!foo(f)` which is short for `>foo f
foo>`. This is dual to the `@` helper, `@foo(f)`, which is short for
`foo> f >foo`.
- Made `List.cons` return a list again since it works nicely with method
overloading. To get a `List+` instead use `List.cons+`.