Add rules to check compiler against origin/main (#347)
Adds `make check-origin` and `make check-origin-fast`. These will use
the mirth compiler in `origin/main` in order to check the current
compiler. This will ensure that we commit features into main before
relying on them in the compiler. (We can still use those new features in
tests and in the examples.)