Introduce conditional compilation via max-mirth-revision and min-mirth-revision. (#362)
This PR introduces `mirth-revision`, a (manually updated) integer that
can be used for conditional compilation across versions of the compiler
via `min-mirth-revision` and `max-mirth-revision` declarations.

The current revision is `20250126001`. When we add some new language
feature we can update this number, let's say we updated it to
`20250126002`, then we can use:

```
min-mirth-revision 20250126002 { ... }
```

to start using the new feature right away inside this block. We can also
use `max-mirth-revision` to target the previous versions:

```
max-mirth-revision 20250126001 { ... }
```

This should make it easier to support multiple versions of the compiler
with the same code, which is useful for bootstrapping (& in general for
libraries).