Improve type error messages
I keep getting type errors and every time I lose a lot of time to just
trying to figure out what type is actually on the stack that is failing
to match. This includes the target namespace of the failed method in the
error message, and also compresses many instances of the same error into
a shorter list to make it more readable.