Use strings instead of input for lexer
The language server is going to need to be able to lex code that is
stored in memory and not in a file (since you can do syntax
highlighting, etc on unsaved files). Given that in most cases it's
better performance to load a string into memory than to operate in a
streaming fashion, it seems like this switchover is overall an
improvement.

It was nice how easy this was to do. The consuming code doesn't need to
aware of the change at all!