Add `+Map(k,+v)` resource
I want to store a `+Map(Str,+Str)` in `+Mirth` that maps file paths to
file contents so the content at a given path can be updated by the
language server.

It's interesting to me how diffent a resource map ends up being vs a
normal one. I had to rethink exactly what the operations would be and am
not sure I chose the right ones or if the naming convention I used is
correct.